Extracts Midpoint and PlacesRating, adds interpolation to Whatsapp links
